Rework events organization
- Swap `receive` system between ServerEventsPlugin and `ClientEventsPlugin` to properly separate what what runs on client or server.
- Move `ServerEventsPlugin::reset` logic inside `ClientEventsPlugin::reset` because this logic runs on client.
- Move `server::events::event_data` module to `core::event_registry::server_event`.
- Move `client::events::event_data` module to `core::event_registry::client_event`.

No functional changes, except `ServerEventsPlugin` and `ClientEventsPlugin` can be disabled on client-only and server-only apps respectively.

Closes #276.
