+++ src/core/event_registry/server_event.rs
@@ -5,17 +5,23 @@ use std::{
 };
 
 use bevy::{
-    ecs::component::{ComponentId, Components},
+    ecs::{
+        component::{ComponentId, Components},
+        entity::MapEntities,
+    },
     prelude::*,
     ptr::{Ptr, PtrMut},
 };
 use bincode::{DefaultOptions, Options};
 use bytes::Bytes;
+use ordered_multimap::ListOrderedMultimap;
+use serde::{de::DeserializeOwned, Serialize};
 
-use super::{DeserializeFn, SendMode, SerializeFn, ServerEventQueue, ToClients};
+use super::EventRegistry;
 use crate::{
     client::replicon_client::RepliconClient,
     core::{
+        channels::{RepliconChannel, RepliconChannels},
         ctx::{ClientReceiveCtx, ServerSendCtx},
         replicon_tick::RepliconTick,
         ClientId,
@@ -26,10 +32,134 @@ use crate::{
     },
 };
 
+/// An extension trait for [`App`] for creating client events.
+pub trait ServerEventAppExt {
+    /// Registers `E` and [`ToClients<E>`] events.
+    ///
+    /// `E` will be emitted on client after sending [`ToClients<E>`] on the server.
+    /// If [`ClientId::SERVER`] is a recipient of the event, then [`ToClients<E>`] will be drained
+    /// after sending to clients and `E` events will be emitted on the server.
+    ///
+    /// Can be called for already existing regular events, a duplicate registration
+    /// for `E` won't be created.
+    ///
+    /// See also [`Self::add_server_event_with`] and the [corresponding section](../index.html#from-server-to-client)
+    /// from the quick start guide.
+    fn add_server_event<E: Event + Serialize + DeserializeOwned>(
+        &mut self,
+        channel: impl Into<RepliconChannel>,
+    ) -> &mut Self {
+        self.add_server_event_with(channel, default_serialize::<E>, default_deserialize::<E>)
+    }
+
+    /// Same as [`Self::add_server_event`], but additionally maps server entities to client inside the event after receiving.
+    ///
+    /// Always use it for events that contain entities.
+    /// See also [`Self::add_server_event`].
+    fn add_mapped_server_event<E: Event + Serialize + DeserializeOwned + MapEntities>(
+        &mut self,
+        channel: impl Into<RepliconChannel>,
+    ) -> &mut Self {
+        self.add_server_event_with(
+            channel,
+            default_serialize::<E>,
+            default_deserialize_mapped::<E>,
+        )
+    }
+
+    /**
+    Same as [`Self::add_server_event`], but uses the specified functions for serialization and deserialization.
+
+    # Examples
+
+    Register an event with [`Box<dyn Reflect>`]:
+
+    ```
+    use std::io::Cursor;
+
+    use bevy::{
+        prelude::*,
+        reflect::serde::{ReflectSerializer, ReflectDeserializer},
+    };
+    use bevy_replicon::{
+        core::ctx::{ClientReceiveCtx, ServerSendCtx},
+        prelude::*,
+    };
+    use bincode::{DefaultOptions, Options};
+    use serde::de::DeserializeSeed;
+
+    let mut app = App::new();
+    app.add_plugins((MinimalPlugins, RepliconPlugins));
+    app.add_server_event_with(
+        ChannelKind::Ordered,
+        serialize_reflect,
+        deserialize_reflect,
+    );
+
+    fn serialize_reflect(
+        ctx: &mut ServerSendCtx,
+        event: &ReflectEvent,
+        cursor: &mut Cursor<Vec<u8>>,
+    ) -> bincode::Result<()> {
+        let serializer = ReflectSerializer::new(&*event.0, ctx.registry);
+        DefaultOptions::new().serialize_into(cursor, &serializer)
+    }
+
+    fn deserialize_reflect(
+        ctx: &mut ClientReceiveCtx,
+        cursor: &mut Cursor<&[u8]>,
+    ) -> bincode::Result<ReflectEvent> {
+        let mut deserializer = bincode::Deserializer::with_reader(cursor, DefaultOptions::new());
+        let reflect = ReflectDeserializer ::new(ctx.registry).deserialize(&mut deserializer)?;
+        Ok(ReflectEvent(reflect))
+    }
+
+    #[derive(Event)]
+    struct ReflectEvent(Box<dyn Reflect>);
+    ```
+    */
+    fn add_server_event_with<E: Event>(
+        &mut self,
+        channel: impl Into<RepliconChannel>,
+        serialize: SerializeFn<E>,
+        deserialize: DeserializeFn<E>,
+    ) -> &mut Self;
+}
+
+impl ServerEventAppExt for App {
+    fn add_server_event_with<E: Event>(
+        &mut self,
+        channel: impl Into<RepliconChannel>,
+        serialize: SerializeFn<E>,
+        deserialize: DeserializeFn<E>,
+    ) -> &mut Self {
+        self.add_event::<E>()
+            .add_event::<ToClients<E>>()
+            .init_resource::<ServerEventQueue<E>>();
+
+        let channel_id = self
+            .world_mut()
+            .resource_mut::<RepliconChannels>()
+            .create_server_channel(channel.into());
+
+        self.world_mut()
+            .resource_scope(|world, mut event_registry: Mut<EventRegistry>| {
+                event_registry.register_server_event(ServerEvent::new(
+                    world.components(),
+                    channel_id,
+                    serialize,
+                    deserialize,
+                ));
+            });
+
+        self
+    }
+}
+
 /// Type-erased functions and metadata for a registered server event.
 ///
 /// Needed so events of different types can be processed together.
-pub(super) struct ServerEventData {
+pub(crate) struct ServerEvent {
     type_id: TypeId,
     type_name: &'static str,
 
@@ -53,8 +183,8 @@ pub(super) struct ServerEventData {
     deserialize: unsafe fn(),
 }
 
-impl ServerEventData {
-    pub(super) fn new<E: Event>(
+impl ServerEvent {
+    fn new<E: Event>(
         components: &Components,
         channel_id: u8,
         serialize: SerializeFn<E>,
@@ -100,15 +230,15 @@ impl ServerEventData {
         }
     }
 
-    pub(super) fn events_id(&self) -> ComponentId {
+    pub(crate) fn events_id(&self) -> ComponentId {
         self.events_id
     }
 
-    pub(super) fn server_events_id(&self) -> ComponentId {
+    pub(crate) fn server_events_id(&self) -> ComponentId {
         self.server_events_id
     }
 
-    pub(super) fn queue_id(&self) -> ComponentId {
+    pub(crate) fn queue_id(&self) -> ComponentId {
         self.queue_id
     }
 
@@ -118,7 +248,7 @@ impl ServerEventData {
     ///
     /// The caller must ensure that `events` is [`Events<ToClients<E>>`]
     /// and this instance was created for `E`.
-    pub(super) unsafe fn send(
+    pub(crate) unsafe fn send(
         &self,
         ctx: &mut ServerSendCtx,
         server_events: &Ptr,
@@ -134,7 +264,7 @@ impl ServerEventData {
     ///
     /// The caller must ensure that `events` is [`Events<E>`], `queue` is [`ServerEventQueue<E>`],
     /// and this instance was created for `E`.
-    pub(super) unsafe fn receive(
+    pub(crate) unsafe fn receive(
         &self,
         ctx: &mut ClientReceiveCtx,
         events: PtrMut,
@@ -151,7 +281,7 @@ impl ServerEventData {
     ///
     /// The caller must ensure that `events` is [`Events<E>`], `server_events` is [`Events<ToClients<E>>`],
     /// and this instance was created for `E`.
-    pub(super) unsafe fn resend_locally(&self, server_events: PtrMut, events: PtrMut) {
+    pub(crate) unsafe fn resend_locally(&self, server_events: PtrMut, events: PtrMut) {
         (self.resend_locally)(server_events, events);
     }
 
@@ -163,7 +293,7 @@ impl ServerEventData {
     ///
     /// The caller must ensure that `queue` is [`Events<E>`]
     /// and this instance was created for `E`.
-    pub(super) unsafe fn reset(&self, queue: PtrMut) {
+    pub(crate) unsafe fn reset(&self, queue: PtrMut) {
         (self.reset)(queue);
     }
 
@@ -172,7 +302,7 @@ impl ServerEventData {
     /// # Safety
     ///
     /// The caller must ensure that this instance was created for `E`.
-    pub(super) unsafe fn serialize<E: Event>(
+    unsafe fn serialize<E: Event>(
         &self,
         ctx: &mut ServerSendCtx,
         event: &E,
@@ -188,7 +318,7 @@ impl ServerEventData {
     /// # Safety
     ///
     /// The caller must ensure that this instance was created for `E`.
-    pub(super) unsafe fn deserialize<E: Event>(
+    unsafe fn deserialize<E: Event>(
         &self,
         ctx: &mut ClientReceiveCtx,
         cursor: &mut Cursor<&[u8]>,
@@ -209,13 +339,19 @@ impl ServerEventData {
     }
 }
 
+/// Signature of server event serialization functions.
+pub type SerializeFn<E> = fn(&mut ServerSendCtx, &E, &mut Cursor<Vec<u8>>) -> bincode::Result<()>;
+
+/// Signature of server event deserialization functions.
+pub type DeserializeFn<E> = fn(&mut ClientReceiveCtx, &mut Cursor<&[u8]>) -> bincode::Result<E>;
+
 /// Signature of server event sending functions.
 type SendFn =
-    unsafe fn(&ServerEventData, &mut ServerSendCtx, &Ptr, &mut RepliconServer, &ConnectedClients);
+    unsafe fn(&ServerEvent, &mut ServerSendCtx, &Ptr, &mut RepliconServer, &ConnectedClients);
 
 /// Signature of server event receiving functions.
 type ReceiveFn = unsafe fn(
-    &ServerEventData,
+    &ServerEvent,
     &mut ClientReceiveCtx,
     PtrMut,
     PtrMut,
@@ -236,7 +372,7 @@ type ResetFn = unsafe fn(PtrMut);
 /// The caller must ensure that `events` is [`Events<ToClients<E>>`]
 /// and `event_data` was created for `E`.
 unsafe fn send<E: Event>(
-    event_data: &ServerEventData,
+    event_data: &ServerEvent,
     ctx: &mut ServerSendCtx,
     server_events: &Ptr,
     server: &mut RepliconServer,
@@ -259,7 +395,7 @@ unsafe fn send<E: Event>(
 /// The caller must ensure that `events` is [`Events<E>`], `queue` is [`ServerEventQueue<E>`]
 /// and `event_data` was created for `E`.
 unsafe fn receive<E: Event>(
-    event_data: &ServerEventData,
+    event_data: &ServerEvent,
     ctx: &mut ClientReceiveCtx,
     events: PtrMut,
     queue: PtrMut,
@@ -341,7 +477,7 @@ unsafe fn reset<E: Event>(queue: PtrMut) {
 ///
 /// The caller must ensure that `event_data` was created for `E`.
 unsafe fn send_with<E: Event>(
-    event_data: &ServerEventData,
+    event_data: &ServerEvent,
     ctx: &mut ServerSendCtx,
     event: &E,
     mode: &SendMode,
@@ -390,7 +526,7 @@ unsafe fn send_with<E: Event>(
 ///
 /// The caller must ensure that `event_data` was created for `E`.
 unsafe fn serialize_with<E: Event>(
-    event_data: &ServerEventData,
+    event_data: &ServerEvent,
     ctx: &mut ServerSendCtx,
     event: &E,
     client: &ConnectedClient,
@@ -434,7 +570,7 @@ unsafe fn serialize_with<E: Event>(
 /// The caller must ensure that `event_data` was created for `E`.
 unsafe fn deserialize_with<E: Event>(
     ctx: &mut ClientReceiveCtx,
-    event_data: &ServerEventData,
+    event_data: &ServerEvent,
     cursor: &mut Cursor<&[u8]>,
 ) -> bincode::Result<(RepliconTick, E)> {
     let tick = DefaultOptions::new().deserialize_from(&mut *cursor)?;
@@ -454,4 +590,73 @@ impl SerializedMessage {
     fn event_bytes(&self) -> &[u8] {
         &self.bytes[self.tick_size..]
     }
+}
+
+/// An event that will be send to client(s).
+#[derive(Clone, Copy, Debug, Event)]
+pub struct ToClients<T> {
+    pub mode: SendMode,
+    pub event: T,
+}
+
+/// Type of server message sending.
+#[derive(Clone, Copy, Debug)]
+pub enum SendMode {
+    Broadcast,
+    BroadcastExcept(ClientId),
+    Direct(ClientId),
+}
+
+/// Stores all received events from server that arrived earlier then replication message with their tick.
+///
+/// Stores data sorted by ticks and maintains order of arrival.
+/// Needed to ensure that when an event is triggered, all the data that it affects or references already exists.
+#[derive(Resource, Deref, DerefMut)]
+struct ServerEventQueue<T>(ListOrderedMultimap<RepliconTick, T>);
+
+impl<T> ServerEventQueue<T> {
+    /// Pops the next event that is at least as old as the specified replicon tick.
+    fn pop_if_le(&mut self, init_tick: RepliconTick) -> Option<(RepliconTick, T)> {
+        let (tick, _) = self.0.front()?;
+        if *tick > init_tick {
+            return None;
+        }
+        self.0
+            .pop_front()
+            .map(|(tick, event)| (tick.into_owned(), event))
+    }
+}
+
+impl<T> Default for ServerEventQueue<T> {
+    fn default() -> Self {
+        Self(Default::default())
+    }
+}
+
+/// Default event serialization function.
+pub fn default_serialize<E: Event + Serialize>(
+    _ctx: &mut ServerSendCtx,
+    event: &E,
+    cursor: &mut Cursor<Vec<u8>>,
+) -> bincode::Result<()> {
+    DefaultOptions::new().serialize_into(cursor, event)
+}
+
+/// Default event deserialization function.
+pub fn default_deserialize<E: Event + DeserializeOwned>(
+    _ctx: &mut ClientReceiveCtx,
+    cursor: &mut Cursor<&[u8]>,
+) -> bincode::Result<E> {
+    DefaultOptions::new().deserialize_from(cursor)
+}
+
+/// Default event deserialization function.
+pub fn default_deserialize_mapped<E: Event + DeserializeOwned + MapEntities>(
+    ctx: &mut ClientReceiveCtx,
+    cursor: &mut Cursor<&[u8]>,
+) -> bincode::Result<E> {
+    let mut event: E = DefaultOptions::new().deserialize_from(cursor)?;
+    event.map_entities(ctx);
+
+    Ok(event)
 }
